caches: extract mod for modes
