+++ src/request/caches/mod.rs
@@ -1,3 +1,6 @@
+pub mod modes;
+
+use crate::request::caches::modes::LocalCache;
 use crate::request::{Method, Request};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
@@ -5,41 +8,10 @@ use std::collections::HashMap;
 
 pub(crate) const CACHES_ENDPOINT: &str = "/rest/v2/caches";
 
-const DEFAULT_CONCURRENCY_LEVEL: i32 = 1000;
-const DEFAULT_ACQUIRE_TIMEOUT: i32 = 15000;
-
 #[derive(Serialize, Deserialize)]
 enum Cache {
     #[serde(rename = "local-cache")]
     LocalCache(LocalCache),
-}
-
-#[derive(Serialize, Deserialize)]
-struct LocalCache {
-    locking: Locking,
-    statistics: bool,
-}
-
-#[derive(Serialize, Deserialize)]
-struct Locking {
-    #[serde(rename = "concurrency-level")]
-    concurrency_level: i32,
-    #[serde(rename = "acquire-timeout")]
-    acquire_timeout: i32,
-    striping: bool,
-}
-
-impl Default for LocalCache {
-    fn default() -> Self {
-        Self {
-            locking: Locking {
-                concurrency_level: DEFAULT_CONCURRENCY_LEVEL,
-                acquire_timeout: DEFAULT_ACQUIRE_TIMEOUT,
-                striping: false,
-            },
-            statistics: true,
-        }
-    }
 }
 
 #[derive(Debug)]
